Bug 689195: Fix signed/unsigned comparison warning in nsLocalFileUnix.cpp. r=smichaud
UltraBlame original commit: 95a1d14917c4638f3c3d634b9d94000f3848571b

+++ xpcom/io/nsLocalFileUnix.cpp
@@ -1429,7 +1429,7 @@ nsLocalFile::IsExecutable(PRBool *_retval)
             "air",         
             "jar"};        
         nsDependentSubstring ext = Substring(path, dotIdx + 1);
-        for (int i = 0; i < NS_ARRAY_LENGTH(executableExts); i++) {
+        for (size_t i = 0; i < NS_ARRAY_LENGTH(executableExts); i++) {
             if (ext.EqualsASCII(executableExts[i])) {
                 
                 *_retval = PR_TRUE;
